moving all of the story version filed to the one tableStatus file
