--- src/PoCs/deleting.tsx
@@ -1,9 +1,0 @@
-import React, {VoidFunctionComponent } from "react";
-
-export const InstanceDeletion: VoidFunctionComponent  = () => {
-    return (
-        <div>
-            <p style={{ color: "#6A6E73" }}>Deleting</p>
-        </div>
-    );
-};
