+++ src/PoCs/style.css
@@ -11,3 +11,7 @@
     font-size: var(--pf-global--icon--FontSize--md);
     color: var(--pf-global--danger-color--100);
   }
+
+ .mas-m-deleting {
+  color: var(--pf-global--Color--200);
+ }
