+++ src/PoCs/tableStatus.tsx
@@ -1,21 +1,131 @@
-import {  Alert, Button, Spinner, HelperText, HelperTextItem, Flex, FlexItem, Split, SplitItem } from "@patternfly/react-core";
+import { Alert, Button, Spinner, HelperText, HelperTextItem, Flex, FlexItem, Split, SplitItem} from "@patternfly/react-core";
 import React, {VoidFunctionComponent } from "react";
 import CheckCircleIcon from '@patternfly/react-icons/dist/js/icons/check-circle-icon';
 import ExclamationCircleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';
 import "./style.css";
 
+type KafkaStatus =
+  | "ready"
+  | "creating" 
+  | "creatingWarning" 
+  | "creatingError"
+  | "failed"
+  | "deleting"
 
-export const ReadyForUse: VoidFunctionComponent  = () => {
-    return (
-        <div>       
-            <Split hasGutter className="mas-c-status">
+
+type StatusLabelProps = {
+  value: KafkaStatus;
+};
+
+export const StatusLabel: VoidFunctionComponent<StatusLabelProps> = ({
+    value
+  }) => {
+
+    switch (value) {
+        case "ready":
+            return (
+                <div>       
+                    <Split hasGutter className="mas-c-status">
+                        <SplitItem>
+                            <CheckCircleIcon className="mas-m-ready"  />      
+                        </SplitItem>
+                        <SplitItem>
+                            Ready
+                        </SplitItem>   
+                    </Split>   
+                </div>
+            );
+
+        case "creating":
+            return (
+                <div>
+                <Split hasGutter className="mas-c-status">
                 <SplitItem>
-                    <CheckCircleIcon className="mas-m-ready"  />      
+                    <Spinner size="md" />
                 </SplitItem>
                 <SplitItem>
-                    Ready
-                </SplitItem>   
+                    <Button variant="link" isInline>
+                        Creating
+                    </Button>
+                    <Flex>
+                            <FlexItem>
+                                <HelperText>
+                                <HelperTextItem variant="indeterminate">This will be ready shortly.</HelperTextItem>
+                                </HelperText> 
+                            </FlexItem>  
+                        </Flex>  
+                </SplitItem>
+                </Split>   
+                </div>
+            );
+            
+        case "creatingWarning":
+            return (
+                <div>
+            <Split hasGutter className="mas-c-status">
+                <SplitItem>
+                    <Spinner size="md" />
+                </SplitItem>
+                <SplitItem>
+                    <Button variant="link" isInline>
+                        Creating
+                    </Button>
+                </SplitItem>
             </Split>   
+            <Alert variant="warning" isInline isPlain title="This is taking longer than expected." />
         </div>
-    );
-};
+        );
+
+        case "creatingError":
+            return (
+                <div>      
+                <Split hasGutter className="mas-c-status">
+                    <SplitItem>
+                        <Spinner size="md" />
+                    </SplitItem>
+                    <SplitItem>
+                        <Button variant="link" isInline>
+                            Creating
+                        </Button>
+                    </SplitItem>
+                </Split>     
+                <Alert variant="danger" isInline isPlain title="This is taking longer than expected." />                 
+            </div>
+            );
+
+        case "failed":
+            return (
+                <div>       
+                    <Split hasGutter className="mas-c-status">
+                        <SplitItem>
+                            <ExclamationCircleIcon className="mas-m-failed"/>      
+                        </SplitItem>
+                        <SplitItem>
+                            Failed
+                        </SplitItem>   
+                    </Split>      
+                </div>
+            );
+                                                  
+        case "deleting":
+            return (
+                <div>
+                    <p className="mas-m-deleting">  Deleting</p>
+                </div>
+            );
+                
+        default:
+            return (
+                <div>       
+                    <Split hasGutter className="mas-c-status">
+                        <SplitItem>
+                            <CheckCircleIcon className="mas-m-ready"  />      
+                        </SplitItem>
+                        <SplitItem>
+                            Ready
+                        </SplitItem>   
+                    </Split>   
+                </div>
+            );
+        }
+    };
