--- src/poc/StatusLabel.css
@@ -1,7 +1,0 @@
-.appservices_kafka-instance-status_icon--completed {
-    color: var(--pf-global--success-color--100);
-  }
-  
-  .appservices_kafka-instance-status_icon--failed {
-    color: var(--pf-global--danger-color--100);
-  }
